move redshift credentials testing to redshift loading location
